Update variables.sqf
Removed "Land_A_tent" from dayz_updateObjects array, and replaced it with "TentStorage".   While techniqually it's the root class of all the tents, it is also the root class of so many other things.   Tentstorage is the class for use with any "tent"... the origional, and both the domed tents are part of this class.